+++ sdk/release_notes.md
@@ -4,7 +4,7 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker.Sdk 1.11.0 (meta package)
+### Microsoft.Azure.Functions.Worker.Sdk <version> (meta package)
 
 - Update extension build project to reference Microsoft.NET.Sdk.Functions 4.2.0
 
@@ -12,7 +12,7 @@
 
 - <entry>
 
-### Microsoft.Azure.Functions.Worker.Sdk.Generators 1.1.0-preview3
+### Microsoft.Azure.Functions.Worker.Sdk.Generators <version>
 
-- Add parsing and support for retry options (#1548).
+- Add support for retry options (#1548)
 - Bug fix for when DefaultValue is not present on an IsBatched prop (#1602).